implementacion de foto 360, inicio page historia, slider reciclado
