+++ turispapa/package.json
@@ -25,6 +25,7 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "postcss": "^8.4.27",
     "react-icons": "^4.10.1",
+    "react-photo-sphere-viewer": "^3.3.5-psv5.1.4",
     "react-router-dom": "^6.14.2",
     "swiper": "^10.2.0",
     "tailwindcss": "^3.3.3",
