+++ turispapa/src/Components/Historia/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function HIstoria() {
+  return (
+    <div>HIstoria</div>
+  )
+}
+
+export default HIstoria
