--- turispapa/src/Components/home/Slider/style.css
@@ -1,69 +1,0 @@
-* {
-    --swiper-navigation-size: 20px;
-    --swiper-navigation-top-offset: 50%;
-    --swiper-navigation-sides-offset: 10px;
-    --swiper-navigation-color: #ffff;
-    --swiper-pagination-color: #ffff;
-}
-
-swiper-container {
-    width: 100%;
-    height: 300px;
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: 8px;
-}
-
-swiper-slide {
-    background-size: cover;
-    background-position: center;
-}
-
-.mySwiper {
-    height: 80%;
-    width: 100%;
-}
-
-.mySwiper2 {
-    height: 20%;
-    box-sizing: border-box;
-    padding: 10px 0;
-}
-
-.mySwiper2 swiper-slide {
-    width: 25%;
-    height: 100%;
-    opacity: 0.4;
-}
-
-.mySwiper2 .swiper-slide-thumb-active {
-    opacity: 1;
-}
-
-swiper-slide img {
-    display: block;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-swiper-container {
-    width: 100%;
-    height: 100%;
-}
-
-swiper-slide {
-    text-align: center;
-    font-size: 18px;
-    background: #fff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-swiper-slide img {
-    display: block;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
