+++ turispapa/src/pages/Home/index.jsx
@@ -1,6 +1,6 @@
 import Experiencias from "../../Components/home/Experiencias"
 import MapaIndex from "../../Components/home/MapaIndex"
-import Slider from "../../Components/home/Slider"
+import Slider from "../../Components/Slider"
 
 /**
  * La función Inicio devuelve un elemento JSX que incluye un encabezado, un componente llamado
@@ -11,8 +11,11 @@ function Home() {
         <>
             <h1 className="mb-6 text-2xl font-semibold">¡La Unión, el tesoro papero!</h1>
             <Experiencias />
-            <Slider/>
-            <MapaIndex/>
+            <h2 className=' text-center m-4 font-semibold text-lg '>50 Experiencias para vivir en La Unión</h2>
+            <Slider
+            width={'w-[80%]'}
+            />
+            <MapaIndex />
         </>
     )
 }
